+++ packages/derivatives-consumer/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.76](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/derivatives-consumer@0.0.75...@injectivelabs/derivatives-consumer@0.0.76) (2021-12-09)
+
+**Note:** Version bump only for package @injectivelabs/derivatives-consumer
+
+
+
+
+
 ## [0.0.75](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/derivatives-consumer@0.0.73...@injectivelabs/derivatives-consumer@0.0.75) (2021-12-07)